+++ DataFinder/src/client/DataFinder/helpers/ParticipantGroup.test.ts
@@ -1,2 +1,2 @@
-
+import * as ParticipantGroup from './ParticipantGroup'
